backend/requirements.txt
APScheduler==3.6.3
channels==1.4.0
daphne==1.4.0
Django==1.11.29
django-apscheduler==0.3.0
django-cors-headers==2.5.3
djangorestframework==3.9.4
mysqlclient==1.4.6
pytz==2020.1
six==1.15.0
sqlparse==0.2.4
tzlocal==2.1
